requirements.txt
groq==0.13.0
streamlit==1.43.2
flask==3.1.0
flask-cors==5.0.1
requests==2.32.3
python-dotenv==1.1.0
gunicorn==23.0.0